1. Change css style after hovering on buttons 2. Add "contiue with faceook" in "login" page
