+++ src/app/auth/login.component.html
@@ -20,6 +20,11 @@
 			<p>Please check your inbox for verification link to verify your email</p>
 		</span>
 	</div>-->
+	
+
+	<div class="facebook" (click)="LoginWithFacebook()">CONTINUE WITH FACEBOOK</div>
+
+	<div class="or-divider"><span>OR</span></div>
 
 	<form #loginForm>
 		<div class="form-group">
@@ -31,7 +36,7 @@
 		</div>
 
 		<div class="form-group">
-			<button class="submit span_69" (click)="LoginWithEmail(email,password); false">Login</button>
+			<button class="submit span_69" (click)="LoginWithEmail(email,password); false">LOGIN</button>
 		</div>
 
 	</form>
