+++ src/app/auth/register.component.css
@@ -44,6 +44,10 @@
 	-moz-border-radius: 5px;
 	border-radius: 5px;
 }
+#form-wrapper .facebook:hover {
+	background-color:#29487d;
+	cursor: pointer;
+}
 
 #form-wrapper .login {
 	-webkit-border-radius: 5px;
@@ -59,6 +63,7 @@
 }
 #form-wrapper .login:hover {
 	background-color: lightgray;
+	cursor: pointer;
 }
 
 #errors {
@@ -87,6 +92,7 @@ form .submit {
 }
 form .submit:hover {
 	background-color: teal;
+	cursor: pointer;
 }
